move partition_key_range
