+++ ydb/core/persqueue/pqtablet/partition/ya.make
@@ -31,7 +31,6 @@ PEERDIR(
     ydb/core/persqueue/common/proxy
     ydb/core/persqueue/public/counters
     ydb/core/persqueue/public/write_meta
-    ydb/core/persqueue/partition_key_range
     ydb/core/persqueue/pqtablet/blob
     ydb/core/persqueue/pqtablet/cache
     ydb/core/persqueue/pqtablet/common
