+++ ydb/core/persqueue/public/partition_key_range/partition_key_range_sequence.h
